Upgrade Artie Transfer (#410)

+++ lib/debezium/converters/money.go
@@ -1,53 +1,53 @@
 package converters
 
 import (
-	"fmt"
-	"github.com/artie-labs/transfer/lib/debezium"
-	"strings"
+    "fmt"
+    "github.com/artie-labs/transfer/lib/debezium"
+    "strings"
 )
 
-const defaultScale = 2
+const defaultScale = uint16(2)
 
 type MoneyConverter struct {
-	// All of these configs are optional
+    // All of these configs are optional
 
-	StripCommas    bool
-	CurrencySymbol string
-	ScaleOverride  *int
+    StripCommas    bool
+    CurrencySymbol string
+    ScaleOverride  *uint16
 }
 
-func (m MoneyConverter) Scale() int {
-	if m.ScaleOverride != nil {
-		return *m.ScaleOverride
-	}
+func (m MoneyConverter) Scale() uint16 {
+    if m.ScaleOverride != nil {
+        return *m.ScaleOverride
+    }
 
-	return defaultScale
+    return defaultScale
 }
 
 func (m MoneyConverter) ToField(name string) debezium.Field {
-	return debezium.Field{
-		FieldName:    name,
-		Type:         debezium.Bytes,
-		DebeziumType: debezium.KafkaDecimalType,
-		Parameters: map[string]any{
-			"scale": fmt.Sprint(m.Scale()),
-		},
-	}
+    return debezium.Field{
+        FieldName:    name,
+        Type:         debezium.Bytes,
+        DebeziumType: debezium.KafkaDecimalType,
+        Parameters: map[string]any{
+            "scale": fmt.Sprint(m.Scale()),
+        },
+    }
 }
 
 func (m MoneyConverter) Convert(value any) (any, error) {
-	valString, isOk := value.(string)
-	if !isOk {
-		return nil, fmt.Errorf("expected string got %T with value: %v", value, value)
-	}
+    valString, isOk := value.(string)
+    if !isOk {
+        return nil, fmt.Errorf("expected string got %T with value: %v", value, value)
+    }
 
-	if m.CurrencySymbol != "" {
-		valString = strings.Trim(valString, m.CurrencySymbol)
-	}
+    if m.CurrencySymbol != "" {
+        valString = strings.Trim(valString, m.CurrencySymbol)
+    }
 
-	if m.StripCommas {
-		valString = strings.ReplaceAll(valString, ",", "")
-	}
+    if m.StripCommas {
+        valString = strings.ReplaceAll(valString, ",", "")
+    }
 
-	return debezium.EncodeDecimal(valString, m.Scale())
+    return debezium.EncodeDecimal(valString, m.Scale())
 }
